+++ components/portal/misc/app_header_options_controller_spec.js
@@ -20,14 +20,12 @@
 /* eslint-env node */
 /* global inject */
 define(['angular-mocks', 'portal'], function() {
-    describe('PortalMainController', function() {
+    describe('AppHeaderOptionsController', function() {
         var $scope;
-        var $localStorage;
         var $document;
         var mainService;
         var NAMES;
 
-        // eslint-disable-next-line no-unused-vars
         var controller;
 
         beforeEach(function() {
@@ -35,10 +33,9 @@ define(['angular-mocks', 'portal'], function() {
         });
 
         beforeEach(inject(function(
-          _$rootScope_, $controller, _$localStorage_, _$document_,
+          _$rootScope_, $controller, _$document_,
           _mainService_, _NAMES_) {
           $scope = _$rootScope_.$new();
-          $localStorage = _$localStorage_;
           $document = _$document_;
           mainService = _mainService_;
           NAMES = _NAMES_;
@@ -46,9 +43,7 @@ define(['angular-mocks', 'portal'], function() {
           NAMES.title = 'AppName';
           $scope.portal.theme.title = 'PortalName';
 
-          controller = $controller('PortalMainController', {
-            '$localStorage': $localStorage,
-            '$scope': $scope,
+          controller = $controller('AppHeaderOptionsController', {
             '$rootScope': _$rootScope_,
             '$document': $document,
             'mainService': mainService,
@@ -56,27 +51,27 @@ define(['angular-mocks', 'portal'], function() {
           });
         }));
 
-        it('should set storage in scope', function() {
-            expect($scope.$storage).not.toBeNull();
-        });
-
-        it('should have an app name defined', function() {
-            expect($scope.NAMES.title).not.toBeNull();
-        });
-
-        it('should set document title to reflect portal name in theme',
+        it('should expose an updateWindowTitle function',
             function() {
+               expect(controller.updateWindowTitle).not.toBeNull();
+        });
+
+        it('should set document title reflecting page title',
+            function() {
+            controller.updateWindowTitle('SpecificPageName');
             // this test is problematic in that it relies on too much
-            // knowledge of what mainService will tell PortalMainController
-            // the window title ought to be. It would be better if this tested
-            // 1. that PortalMainController called mainService with the correct
+            // knowledge of what mainService will tell
+            // AppHeaderOptionsController the window title ought to be. It
+            // would be better if this tested
+            // 1. that the controller called mainService with the correct
             // arguments, and
-            // 2. that PortalMainController set the window title to whatever
+            // 2. that the controller set the window title to whatever
             // mainService told it the title ought to be.
             // that is, mock mainService both to monitor arguments
-            // PortalMainController calls it with and to fake its response to
-            // see that PortalMainController honors it
-            expect($document[0].title).toEqual('AppName | PortalName');
+            // the controller calls it with and to fake its response to
+            // see that the controller honors it
+            expect($document[0].title)
+              .toEqual('SpecificPageName | AppName | PortalName');
         });
     });
 });